refactor: Remove no-longer required OldPredicate type
